pyproject.toml
[project]
name = "rocketwatch"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.13.5",
    "anthropic>=0.52.0,<1.0.0",
    "google-genai>=1.0.0,<2.0.0",
    "openai>=1.0.0,<3.0.0",
    "aiocache==0.12.3",
    "anyascii==0.3.3",
    "beautifulsoup4==4.14.3",
    "bidict==0.23.1",
    "cachetools==7.0.5",
    "colorama==0.4.6",
    "cronitor==4.9.0",
    "dice==4.0.0",
    "discord.py[voice]==2.7.1",
    "discord-ext-voice-recv>=0.5.0",
    "pydub>=0.25.1",
    "eth-typing==6.0.0",
    "eth-utils==6.0.0",
    "etherscan_labels @ git+https://github.com/haloooloolo/etherscan-labels",
    "graphql_query==1.4.0",
    "hexbytes==1.3.1",
    "humanize==4.15.0",
    "matplotlib==3.10.8",
    "numpy==2.4.4",
    "pillow==12.2.0",
    "psutil==7.2.2",
    "pydantic>=2.0.0,<3.0.0",
    "pymongo==4.16.0",
    "pytz==2026.1.post1",
    "regex==2026.3.32",
    "retry-async==0.1.4",
    "seaborn==0.13.2",
    "tabulate==0.10.0",
    "termplotlib==0.3.9",
    "uptime==3.0.1",
    "web3>=7.0.0,<8.0.0",
]


[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore:websockets.legacy is deprecated:DeprecationWarning",
]

[tool.mypy]
python_version = "3.12"
explicit_package_bases = true
strict = true
disallow_untyped_calls = false
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = [
    "aiocache", "aiocache.*",
    "ens", "ens.*",
    "etherscan_labels", "etherscan_labels.*",
    "retry_async", "retry_async.*",
    "termplotlib", "termplotlib.*",
    "cronitor", "cronitor.*",
    "dice", "dice.*",
    "uptime", "uptime.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "RUF", "SIM", "UP", "W"]
ignore = ["E501", "E203", "E231"]

[tool.ruff.lint.isort]
known-first-party = ["rocketwatch", "sentinel"]

[dependency-groups]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "ruff>=0.15.9",
    "types-cachetools>=6.2.0.20260317",
    "types-colorama>=0.4.15.20250801",
    "types-psutil>=7.2.2.20260402",
    "types-pytz>=2026.1.1.20260304",
    "types-regex>=2026.3.32.20260329",
    "types-seaborn>=0.13.2.20260402",
    "types-tabulate>=0.10.0.20260308",
]
